fix: 공통함수를 모듈로 따로 관리하게 변경
