+++ crawling_python/carawling_saramin.py
@@ -1,43 +1,31 @@
-import os
-import time
-
-from selenium import webdriver
+from global_utils import *
 from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import WebDriverWait
 from selenium.webdriver.support import expected_conditions as EC
 from datetime import datetime, timedelta
 from dotenv import load_dotenv
 import re
-import mysql.connector
 
 load_dotenv()
 
-db = mysql.connector.connect(
-    host="localhost",
-    port=3307,
-    user=os.getenv('DB_USERNAME'),
-    password=os.getenv('DB_USER_PASSWORD'),
-    database=os.getenv('DB_DATABASE')
-)
 
-cursor = db.cursor()
-
-driver = webdriver.Chrome()
-
-page_number = 1
-while True:
+def crawling_job_data(driver, page_number):
     url = f'https://www.saramin.co.kr/zf_user/jobs/public/list?page={page_number}&isAjaxRequest=y&cat_kewd=84%2C92%2C87%2C86&company_cd=0%2C1%2C2%2C3%2C4%2C5%2C6%2C7%2C9%2C10&panel_type=domestic&search_optional_item=n&search_done=y&panel_count=y&preview=y'
     driver.get(url)
 
     wait = WebDriverWait(driver, 10)
 
     try:
-        companies = wait.until(EC.presence_of_all_elements_located((By.CSS_SELECTOR, '.list_item .col.company_nm a.str_tit, .list_item .col.company_nm span.str_tit')))
-        contents = wait.until(EC.presence_of_all_elements_located((By.CSS_SELECTOR, '.list_body .col.notification_info .job_tit .str_tit')))
-        urls = wait.until(EC.presence_of_all_elements_located((By.CSS_SELECTOR, '.list_body .col.notification_info .job_tit a.str_tit')))
-        dates = wait.until(EC.presence_of_all_elements_located((By.CSS_SELECTOR, '.list_body .col.support_info .support_detail .date')))
+        companies = wait.until(EC.presence_of_all_elements_located(
+            (By.CSS_SELECTOR, '.list_item .col.company_nm a.str_tit, .list_item .col.company_nm span.str_tit')))
+        contents = wait.until(EC.presence_of_all_elements_located(
+            (By.CSS_SELECTOR, '.list_body .col.notification_info .job_tit .str_tit')))
+        urls = wait.until(EC.presence_of_all_elements_located(
+            (By.CSS_SELECTOR, '.list_body .col.notification_info .job_tit a.str_tit')))
+        dates = wait.until(EC.presence_of_all_elements_located(
+            (By.CSS_SELECTOR, '.list_body .col.support_info .support_detail .date')))
     except:
-        break
+        return None
 
     data_list = []
 
@@ -64,15 +52,30 @@ while True:
 
         data_list.append((company, content, url, date))
 
-    insert_query = "INSERT INTO job (company_name, title, url, expiration_date) VALUES (%s, %s, %s, %s)"
-    cursor.executemany(insert_query, data_list)
+    return data_list
 
-    db.commit()
 
-    page_number += 1
+def main():
+    driver = get_driver()
+    page_number = 1
 
-cursor.close()
+    while True:
+        job_data = crawling_job_data(driver, page_number)
+        if job_data is None:
+            break
 
-db.close()
+        db = get_database_connect()
+        cursor = db.cursor()
+        insert_query = "INSERT INTO job (company_name, title, url, expiration_date) VALUES (%s, %s, %s, %s)"
+        cursor.executemany(insert_query, job_data)
+        db.commit()
+        cursor.close()
+        db.close()
 
-driver.quit()
+        page_number += 1
+
+    driver.quit()
+
+
+if __name__ == "__main__":
+    main()
